terminei de duplicar e organizar
