+++ lib/HomeNormal/Encartes.dart
@@ -1,15 +1,15 @@
 import 'dart:async';
+import 'package:encarte_facil_2/DesignSystem/Components/Cell%20Encarte.dart';
+import 'package:encarte_facil_2/DesignSystem/Components/ReloadButton.dart';
+import 'package:encarte_facil_2/DesignSystem/DesignTokens.dart';
+import 'package:encarte_facil_2/HomeNormal/Home.dart';
+import 'package:encarte_facil_2/HomeNormal/ProdutosEncarte.dart';
+import 'package:encarte_facil_2/Logic/controller.dart';
+import 'package:encarte_facil_2/Model/Produto.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:provider/provider.dart';
-import 'DesignSystem/Components/Cell Encarte.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
-import 'DesignSystem/Components/ReloadButton.dart';
-import 'DesignSystem/DesignTokens.dart';
-import 'Logic/controller.dart';
-import 'Model/Produto.dart';
-import 'ProdutosEncarte.dart';
-
 
 class Encartes extends StatefulWidget {
   const Encartes({key}) : super(key: key);
@@ -94,7 +94,19 @@ class _EncartesState extends State<Encartes> {
                                       ),
                                     ),
                                     Spacer(),
-                                    ReloadButtonWidget(controller)
+                                    GestureDetector(
+                                      child: ReloadButtonWidget(controller),
+                                      onTap: () {
+                                        Navigator.pushReplacement(
+                                          context,
+                                          PageRouteBuilder(
+                                            pageBuilder: (context, animation1, animation2) => HomeWidget(),
+                                            transitionDuration: Duration.zero,
+                                            reverseTransitionDuration: Duration.zero,
+                                          ),
+                                        );
+                                      },
+                                    )
                                   ],
                                 ),
                                 Padding(
