+++ lib/HomeNormal/main.dart
@@ -1,12 +1,10 @@
-import 'package:encarte_facil_2/Nova%20Home/Home.dart';
-import 'package:encarte_facil_2/Premium%20flow/InformaCodigoPRO.dart';
+import 'package:encarte_facil_2/HomeNormal/Home.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter_uxcam/flutter_uxcam.dart';
 import 'package:provider/provider.dart';
-import 'Logic/Functions.dart';
-import 'Logic/controller.dart';
+import '../Logic/controller.dart';
 
 void main() async{
   WidgetsFlutterBinding.ensureInitialized();
